minor cleanup

+++ mcxr-play/src/main/java/net/sorenon/mcxr/play/openxr/OpenXRSwapchain.java
@@ -62,10 +62,14 @@ public class OpenXRSwapchain implements AutoCloseable {
     @Override
     public void close() {
         XR10.xrDestroySwapchain(handle);
-        images.close();
-        RenderSystem.recordRenderCall(() -> {
-            innerFramebuffer.delete();
-            framebuffer.delete();
-        });
+        if (images != null) {
+            images.close();
+        }
+        if (framebuffer != null) {
+            RenderSystem.recordRenderCall(() -> {
+                innerFramebuffer.delete();
+                framebuffer.delete();
+            });
+        }
     }
 }
